Gtest: Fix sign vs unsiged compile warning
The number in C++ is signed int, when compared with .size() that returns
size_t it raises signg vs unsinged warning

The C++ std::strlen returns size_t but the regex smatch .length()
returns difference_type_t which is of ssize_t. Use explicic casts as all
std::strlen of ststic string are lower then size_t::max.

The stat's mode function return unsiged int, use it in test as well

The gtest library has some unused constant values, turn off that warning
as it creates compile time wrarning.

+++ src/modules/complianceengine/tests/OptionalTest.cpp
@@ -113,7 +113,7 @@ TEST_F(OptionalTest, ArrowOperator)
 {
     auto opt = Optional<std::string>("foo");
     ASSERT_TRUE(opt.HasValue());
-    ASSERT_EQ(opt->size(), 3);
+    ASSERT_EQ(opt->size(), (size_t)3);
     opt->append("bar");
-    ASSERT_EQ(opt->size(), 6);
+    ASSERT_EQ(opt->size(), (size_t)6);
 }
